requirements.txt
asgiref==3.9.1
Django==5.2.6
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
sqlparse==0.5.3
typing_extensions==4.15.0
